requirements.txt
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.8.1
billiard==4.2.1
celery==5.5.2
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.0
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
contourpy==1.3.1
cycler==0.12.1
dj-rest-auth==7.0.1
Django==5.2
django-cors-headers==4.7.0
django-filter==24.3
django-redis==5.4.0
django_celery_results==2.6.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
fonttools==4.56.0
gunicorn==23.0.0
idna==3.10
kiwisolver==1.4.8
kombu==5.5.3
Markdown==3.7
matplotlib==3.10.0
mpmath==1.3.0
mysql==0.0.3
mysql-connector-python==9.3.0
mysqlclient==2.2.7
networkx==3.4.2
numpy==2.2.2
packaging==24.2
pillow==11.2.1
prompt_toolkit==3.0.51
psycopg2-binary==2.9.10
pydantic==2.11.4
pydantic_core==2.33.2
PyJWT==2.9.0
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==6.1.0
requests==2.32.3
six==1.17.0
sqlparse==0.5.3
sympy==1.13.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
vine==5.1.0
wcwidth==0.2.13
